+++ CHANGELOG.md
@@ -14,3 +14,4 @@ master
 * Fix Coveralls tool export
 * Fix badges
 * Disable audit of Log (cf. https://sonata-project.org/bundles/doctrine-orm-admin/master/doc/reference/audit.html)
+* Remove deprecation about configuration tree builder without a root node